requirements.txt
fastapi==0.129.0
mangum==0.21.0
boto3==1.42.59
bcrypt==5.0.0
PyJWT==2.11.0
pydantic[email]==2.12.5
python-dotenv==1.2.1
uvicorn==0.40.0
email-validator==2.3.0
cryptography==46.0.5
python-multipart==0.0.20